requirements.txt
openai>=1.40.0
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
feedparser>=6.0.11
python-dotenv>=1.0.1
jinja2>=3.1.4
PyYAML>=6.0.2
